Finalizing LDAP sync
